Arm backend: Initial quantization support for conditional
The standard prepare/convert_pt2 does not seem to support
quantization out of the box. Instead, a quantization call
is introduced in the TOSAQuantizer, that does the neccessary
steps to get correct quantization on submodules. A custom
Quantize step is needed in the ArmTester to make this work
in testing.

Additionally, getting correct quantization parameters needs
some delicate handling. The model is calibrated twice,
once for each code path. Because of this, the observers outside
the if/else submodules see different data than the observers
inside the submodules. Rescales need to be inserted to handle
this. To get a correctly traced graph at all times, we
  1. Fold the outmost quant ops in the submodules at the same
     time as the cond is folded. Add qparam meta to folded
     nodes inside submodule.
  2. Use this meta in the InsertCondRescale pass to
     insert a tosa.RESCALE to handle the different qparams.
  3. After this, the submodule's q-dq nodes can be folded normally.

Signed-off-by: Erik Lundell <erik.lundell@arm.com>
Change-Id: Ia37c4478a661d1156f537ca6d4e1e17f0f130d24